+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-04-10 20:49:17
-// Commit: chore(core): update dependencies and cleanup unused imports in core
+// Updated at: 2026-04-14 17:00:13
+// Commit: fix(core): correct layout constraints and overflow in core
 
 class coreComponentState {
-  final String commitRef = '301';
-  final String lastMessage = 'chore(core): update dependencies and cleanup unused imports in core';
-  final DateTime timestamp = DateTime.parse('2026-04-10T20:49:17+05:00');
+  final String commitRef = '371';
+  final String lastMessage = 'fix(core): correct layout constraints and overflow in core';
+  final DateTime timestamp = DateTime.parse('2026-04-14T17:00:13+05:00');
 }